requirements.txt
# Dependências principais para o RAG Chatbot ICTA
# Instalação: pip install -r requirements.txt

# Busca vetorial e embeddings
faiss-cpu>=1.7.4
sentence-transformers>=2.2.2
numpy>=1.21.0

# Modelos de linguagem e transformers
transformers>=4.21.0
torch>=1.12.0
accelerate>=0.20.0
huggingface_hub>=0.16.0

# Utilitários e interface
tqdm>=4.64.0
colorama>=0.4.5
requests>=2.28.0

# Dependências opcionais para funcionalidades avançadas (comentadas por padrão)
# langchain>=0.0.300
# langchain-community>=0.0.20

# Dependências para desenvolvimento e testes
pytest>=7.0.0
pytest-cov>=4.0.0
black>=22.0.0
flake8>=5.0.0
mypy>=1.0.0

# Para criação de pacotes
setuptools>=65.0.0
wheel>=0.37.0
